move common test helpers to tests/utils.py
